fix: update to reflect the rename
