+++ tests/ai_hub/mcp_servers/upgrade/conftest.py
@@ -6,12 +6,13 @@ import yaml
 from kubernetes.dynamic import DynamicClient
 from ocp_resources.config_map import ConfigMap
 from ocp_resources.resource import ResourceEditor
-from tests.model_registry.mcp_servers.config.constants import (
+
+from tests.ai_hub.mcp_servers.config.constants import (
     MCP_CATALOG_SOURCE,
     MCP_SERVERS_YAML,
 )
-from tests.model_registry.mcp_servers.config.utils import get_mcp_catalog_sources
-from tests.model_registry.utils import wait_for_mcp_catalog_api, wait_for_model_catalog_pod_ready_after_deletion
+from tests.ai_hub.mcp_servers.config.utils import get_mcp_catalog_sources
+from tests.ai_hub.utils import wait_for_mcp_catalog_api, wait_for_model_catalog_pod_ready_after_deletion
 
 LOGGER = structlog.get_logger(name=__name__)
