+++ tests/ai_hub/mcp_servers/upgrade/test_mcp_catalog_upgrade.py
@@ -4,7 +4,8 @@ import pytest
 import structlog
 import yaml
 from ocp_resources.config_map import ConfigMap
-from tests.model_registry.mcp_servers.config.constants import MCP_CATALOG_SOURCE_ID
+
+from tests.ai_hub.mcp_servers.config.constants import MCP_CATALOG_SOURCE_ID
 
 LOGGER = structlog.get_logger(name=__name__)
